fix: move getSponsorships function from sponsor forms file into summit actions file
