+++ src/reducers/sponsors/show-pages-list-reducer.js
@@ -22,8 +22,10 @@ import {
   SHOW_PAGE_DELETED,
   RESET_SHOW_PAGE_FORM
 } from "../../actions/show-pages-actions";
-import { SET_CURRENT_SUMMIT } from "../../actions/summit-actions";
-import { RECEIVE_GLOBAL_SPONSORSHIPS } from "../../actions/sponsor-forms-actions";
+import {
+  SET_CURRENT_SUMMIT,
+  RECEIVE_GLOBAL_SPONSORSHIPS
+} from "../../actions/summit-actions";
 import { denormalizePageModules } from "../../utils/page-template";
 
 const DEFAULT_SHOW_PAGE = {
